+++ src/constants.js
@@ -10,16 +10,12 @@ import 'intl';
 import 'intl/locale-data/jsonp/en';
 
 import { Network } from '@hathor/wallet-lib';
-import { STORE } from './store';
 import {
   _IS_MULTI_TOKEN as IS_MULTI_TOKEN,
   _DEFAULT_TOKEN as DEFAULT_TOKEN,
   _PRIMARY_COLOR as PRIMARY_COLOR,
   _SENTRY_DSN as SENTRY_DSN,
 } from './config';
-
-// Re-export store object.
-export { STORE };
 
 /**
  * This is the environment stage that will be used to load the unleash feature flags.
