+++ src/screens/DecideStackScreen.js
@@ -7,7 +7,7 @@
 
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { STORE } from '../constants';
+import { STORE } from '../store';
 
 
 import { onExceptionCaptured } from '../actions';
